backend/requirements.txt
fastapi
uvicorn
asyncpg
pydantic
python-dotenv
openai
python-jose[cryptography]
httpx